Moved the vertical spin button widget into spin button's mod.rs
