+++ src/widget/mod.rs
@@ -358,10 +358,6 @@ pub mod tooltip {
     }
 }
 
-pub mod vertical_spin_button;
-#[doc(inline)]
-pub use vertical_spin_button::*;
-
 pub mod warning;
 #[doc(inline)]
 pub use warning::*;
